Updated V1.1
Not much has changed, in this version, now it can only return the maximum value from the arguments, but the withdrawal is that it always returns the maximum value (defeating its main purpose).  I'm still trying to figure out how to do it but for now, it returns a better value than before.